requirements.txt
# Python 3.7.x
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Pillow==7.0.0
PyVCF==0.6.8
gffutils==0.10.1
h5py==2.10.0
matplotlib==3.1.1
numexpr==2.7.1
numpy==1.17.2
opencv-python==4.2.0.34
pandas==0.25.1
pyfaidx==0.5.8
pysam==0.16.0.1
requests==2.21.0
scikit-learn==0.24.1
scipy==1.4.1
seaborn==0.9.0
scikit-learn==0.24.1
statsmodels==0.11.1
tables==3.6.1
tensorboard==2.1.1
tensorflow-estimator==2.1.0
torch==1.7.1+cu110
torchvision==0.8.2+cu110
torch-summary==1.1.8
